パッケージ名の修正、go.modの最新化、カテゴリのミスマッチのハンドリング追加、deleted_atの型を修正、テーブル定義のamountの型を修正、created_atとupdate_atの手動セットを削除（AIレビュー指摘対応）

+++ db/init/01_create_tables.sql
@@ -15,7 +15,7 @@ CREATE TABLE categories (
 CREATE TABLE transactions (
     id               SERIAL PRIMARY KEY,
     title            VARCHAR(100) NOT NULL,
-    amount           INTEGER NOT NULL,                            -- 整数のみ。金額に小数は不要
+    amount           BIGINT NOT NULL,                            -- 整数のみ。金額に小数は不要
     type             VARCHAR(10)  NOT NULL CHECK (type IN ('expense', 'income')), -- 支出 or 収入
     category_id      INTEGER NOT NULL,
     transaction_date DATE NOT NULL,                               -- 時分秒を含まない「決済日」
